% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc.SE.rowcluster}
\alias{calc.SE.rowcluster}
\alias{calc.SE.bicluster}
\title{Calculate standard errors of clustering parameters.}
\usage{
calc.SE.rowcluster(long.df, clust.out,
  optim.control = default.optim.control())

calc.SE.bicluster(long.df, clust.out,
  optim.control = default.optim.control())
}
\arguments{
\item{long.df}{The data frame, in long format, as passed to \code{rowclustering}
or \code{biclustering}.}

\item{clust.out}{For \code{calc.SE.rowcluster}, a \code{rowclustering}
object. For \code{calc.SE.bicluster}, a \code{biclustering} object.}

\item{optim.control}{control list for the \code{optim} call within the M step
of the EM algorithm. See the control list Details in the \code{optim}
manual for more info.}
}
\value{
The standard errors corresponding to the elements of \code{clust.out$outvect}.
}
\description{
Calculate SE of parameters fitted using \code{\link{rowclustering}} or
\code{\link{biclustering}}. Cannot currently be applied to
\code{\link{columnclustering}} output.
}
\details{
Calculates SE by running \code{optimHess} (see \code{\link[stats]{optim}}) on
the incomplete-data log-likelihood to find the hessian at the fitted parameter
values from \code{\link{rowclustering}} or \code{\link{biclustering}}.
Then the square roots of the diagonal elements
of the negative inverse of the hessian are the standard errors of the parameters
i.e. \code{SE <- sqrt(diag(solve(-optim.hess))}.

Note that SE values are \strong{only} calculated for the independent parameters
i.e. if the alpha parameters sum to zero, SE values will only be calculated
for the first RG-1 alpha values, etc.

The function requires an input which is the output of \code{\link{rowclustering}},
which includes the component \code{outvect}, the final vector of independent
parameter values from the EM algorithm, which will correspond to a subset of
the parameter values in \code{parlist.out}.
}
\section{Functions}{
\itemize{
\item \code{calc.SE.rowcluster}: SE for rowclustering

\item \code{calc.SE.bicluster}: SE for biclustering
}}

