\name{clustord-package}
\alias{clustord-package}
\alias{clustord}
\docType{package}
\title{
\packageTitle{clustord}
}
\description{
\packageDescription{clustord}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clustord}
\packageIndices{clustord}
The clustord package can perform row, column or biclustering on an ordinal or
binary dataset, using one of three models: the Ordered Stereotype Model (OSM),
the Proportional Odds Model (POM) or the Binary model (Binary).

The three main functions are \code{rowclustering}, \code{columnclustering} and
\code{biclustering}.

There are also subsidiary functions \code{calc.SE.rowcluster} and
\code{calc.SE.bicluster}. The first one calculates the standard error for the
final parameters of rowclustering or columnclustering, and the second one
calculates the standard error for the final parameters of biclustering.
}
\author{
\packageAuthor{clustord}

Maintainer: \packageMaintainer{clustord}
}
\references{
D. Fernández, R. Arnold, S. Pledger, I. Liu, & R. Costilla, “Finite mixture biclustering of discrete type multivariate data,” Advances in Data Analysis and Classification, vol. 13(1), pp. 117–143, 2019.

J. A. Anderson, “Regression and ordered categorical variables,” Journal of the Royal Statistical Society: Series B (Methodological) vol. 46(1), pp. 1–22, 1984.
}
\keyword{ package }
\seealso{
}
\examples{
}
