\name{clustPOM-package}
\alias{clustPOM-package}
\alias{clustPOM}
\docType{package}
\title{
\packageTitle{clustPOM}
}
\description{
\packageDescription{clustPOM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clustPOM}
\packageIndices{clustPOM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{clustPOM}

Maintainer: \packageMaintainer{clustPOM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
