% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.cluster.comparisons}
\alias{calc.cluster.comparisons}
\title{Calculate comparison measures between two sets of clustering results}
\usage{
calc.cluster.comparisons(ppr1, ppr2)
}
\arguments{
\item{ppr1}{Posterior probabilities of cluster membership, calculated as ppr.m
or ppc.m in the output of \code{\link{rowclustering}}, \code{\link{columnclustering}}
or \code{\link{biclustering}}.
The rows of ppr1 give the entries that have been clustered, and each column
corresponds to one cluster.}

\item{ppr2}{Posterior probabilities of cluster membership from a different
clustering run, which will be compared to ppr1.}
}
\value{
A list with components:

    \code{ARI}: Adjusted Rand Index.

    \code{NVI}: Normalised Variation of Information.

    \code{NID}: Normalised Information Distance.
}
\description{
Given two sets of posterior probabilities of membership for clusters,
calculate three measures to compare the clustering memberships.
}
\details{
The three measures are the Adjusted Rand Index (ARI), the Normalised
Variation of Information (NVI) and the Normalised Information Distance (NID).

The three measures are documented in

Fernández, D., & Pledger, S. (2016). Categorising count data into ordinal
responses with application to ecological communities.
Journal of agricultural, biological, and environmental statistics (JABES),
21(2), 348--362.
}
