\name{clustord-package}
\alias{clustord-package}
\alias{clustord}
\docType{package}
\title{
\packageTitle{clustord}
}
\description{
\packageDescription{clustord}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clustord}
\packageIndices{clustord}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{clustord}

Maintainer: \packageMaintainer{clustord}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
