% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{osm}
\alias{osm}
\title{Logistic regression using ordered stereotype model (OSM)}
\usage{
osm(
  formula,
  data,
  weights,
  start,
  ...,
  subset,
  na.action,
  Hess = FALSE,
  model = TRUE
)
}
\arguments{
\item{formula:}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response should be a factor (preferably
an ordered factor), which will be interpreted as an ordinal response, with
levels ordered as in the factor. The model must have an intercept, which
will then be removed because it is replaced by the fitted mu values;
attempts by the user to remove the intercept in advance will lead to a
warning. An offset may be used.
See the documentation of \link[stats]{formula} for other details.}

\item{data:}{an optional data frame, list or environment in which to
interpret the variables occurring in \code{formula}.}

\item{weights:}{optional case weights in fitting. Default to 1.}

\item{start:}{initial values for the parameters. This is in the format
\code{c(beta coefficients, mu, u)}: see the Values section.}

\item{...:}{additional arguments to be passed to \link[stats]{optim}, most
often a \code{control} argument.}

\item{subset:}{expression saying which subset of the rows of the data should
be used in the fit. All observations are included by default.}

\item{na.action:}{a function to filter missing data.}

\item{Hess:}{logical for whether the Hessian (the observed information
matrix) should be returned. Use this if you intend to call \code{summary}
or \code{vcov} on the fit.}

\item{model:}{logical for whether the model matrix should be returned.}

\item{method:}{logistic or probit or (complementary) log-log or cauchit
(corresponding to a Cauchy latent variable).}
}
\value{
A object of class "osm". This has components:
\describe{
  \item{\code{beta}}{the coefficients of the covariates within the linear predictor, which has no intercept.}
  \item{\code{mu}}{the intercepts for the class boundaries.}
  \item{\code{phi}}{the class-specific scoring parameters that scale the effect of the linear predictor.}
  \item{\code{u}}{the raw scoring parameters fitted by \code{optim}. See Details.}
  \item{\code{deviance}}{the residual deviance.}
  \item{\code{fitted.values}}{a matrix, with a column for each level of the response.}
  \item{\code{lev}}{the names of the response levels.}
  \item{\code{terms}}{the \code{terms} structure describing the model.}
  \item{\code{df.residual}}{the number of residual degrees of freedoms, calculated using the weights.}
  \item{\code{edf}}{the (effective) number of degrees of freedom used by the model}
  \item{\code{n, nobs}}{the (effective) number of observations, calculated using the weights. (\code{nobs} is for use by \link[MASS]{stepAIC}).}
  \item{\code{call}}{the matched call.}
  \item{\code{convergence}}{the convergence code returned by \code{optim}.}
  \item{\code{niter}}{the number of function and gradient evaluations used by \code{optim}.}
  \item{\code{eta}}{the linear predictor (including any offset).}
  \item{\code{Hessian}}{(if \code{Hess} is true), the Hessian for the final fit.
  Note that this is a numerical approximation derived from the optimization
  process.}
  \item{\code{model}}{(if \code{model} is true), the model frame used.}
  \item{\code{na.action}}{the function used to filter missing data.}
  \item{\code{xlevels}}{list of factor levels from any factors or character vectors in the \code{terms} of the model.}
}
}
\description{
Fits a logistic regression model to an ordered factor response using the
Ordered Stereotype Models (OSM), using the reparametrization of Fernandez et
al (2016) and enforcing the non-decreasing constraint on \code{phi_k} values.
}
\details{
See Anderson (1984) for the original definition of the ordered stereotype
model, and see Fernandez et al. (2016) for the reparametrization of the
\code{phi} values. Also see Agresti (2010) for more information about the
stereotype model and other ordinal models.

The model assumes that the response is an ordered variable Y with q levels.

This model is a baseline logit model, in which the distributions of levels
k = 2,...,q of the response Y are each defined separately relative to the
baseline. But the model is more parsimonious than the fully flexible baseline
logit model, because the effects of the linear predictor on the logits of
levels k = 2,...,q are proportional to each other.

The model takes the form

log(P(Y = k|x)/P(Y = 1|x)) = mu_k + phi_k*eta for k = 2,...,q,

with \code{eta} being the linear predictor, i.e. \code{beta*x} where
\code{x} are the covariates and \code{beta} are the coefficients. There is no
\code{beta_0} term, because that is absorbed into the class boundary parameters
\code{mu_k}.

This function fits the values of \code{mu_k}, \code{phi_k} and \code{beta}.

\code{mu_1} = 0 for identifiability, so there are \code{q-1} independent
values of \code{mu_k}.

The \code{phi} values are constrained to be non-decreasing, with

phi_1 = 0 <= phi_2 <= ... <= phi_k = 1

so there are \code{q-2} independent but non-decreasing values of
\code{phi_k}. This is the ordered form of the stereotype model.

This is unlike the stereotype models fitted with the package
\url{https://cran.r-project.org/web/packages/ordinalgmifs/index.html}, which
do NOT enforce the ordering constraint on the \code{phi} values, and merely
enforce the 0 and 1 lower and upper limits for \code{phi}.

If the user specifies the starting values for \code{beta}, \code{mu} and
\code{phi}, they can be specified in a vector, in that order. There should be
\code{length(x)} independent values of \code{beta},  \code{q-1} independent
values of \code{mu}, and \code{q-2} independent values of \code{phi}. BUT the
starting values for \code{phi} do not correspond directly to \code{phi},
because \code{phi} is restricted to being increasing and between 0 and 1, so
instead the starting values are treated as elements \code{u[2:q-1]} of a
vector \code{u} which can be between \code{-Inf} and \code{+Inf}, and then

    \code{phi[2] <- expit(u[2])} and

    \code{phi[k] <- expit(u[2] + sum(exp(u[3:k])))} for k between 3 and q-1

    \code{(phi[1] = 0 and phi[q] = 1)}.

The fitted values of \code{phi_k} and \code{u_k} are both included in the output.
}
\references{
Fernandez, D., Arnold, R., & Pledger, S. (2016). Mixture-based clustering for the ordered stereotype model. *Computational Statistics & Data Analysis*, 93, 46-75.
Anderson, J. A. (1984). Regression and ordered categorical variables. *Journal of the Royal Statistical Society: Series B (Methodological)*, 46(1), 1-22.
Agresti, A. (2010). *Analysis of ordinal categorical data* (Vol. 656). John Wiley & Sons.
}
